# backend/docs/swagger.yaml
components:
  schemas:
    data:
      properties:
        data:
          $ref: '#/components/schemas/models.UserResponse'
      type: object
...
    models.DeploymentResponse:
      properties:
        deployment_id:
          type: string
        message:
          type: string
        started_at:
          type: string
        status:
          type: string
      type: object
    models.DeploymentStatus:
      properties:
        assets_url:
          type: string
        build_id:
          type: string
        deployment_id:
          type: string
        last_deployed_at:
          type: string
        last_update_at:
          type: string
        message:
          type: string
        preview_url:
          type: string
        production_url:
          type: string
        shop_id:
          type: string
        status:
          description: '"deployed", "deploying", "failed", "not_deployed"'
          type: string
        subdomain:
          type: string
        template_name:
          type: string
        template_version:
          type: string
      type: object
...
    models.ErrorResponse:
      description: Generic API error response
      properties:
        code:
          type: integer
        errors:
          items:
            $ref: '#/components/schemas/models.Error'
          type: array
          uniqueItems: false
        message:
          type: string
        meta:
          $ref: '#/components/schemas/models.Meta'
        status:
          type: string
      type: object
...
    models.StoreDeploymentRequest:
      properties:
        data_override:
          additionalProperties:
            type: string
          type: object
        shop_id:
          type: string
        subdomain:
          type: string
        template_name:
          type: string
        version:
          type: string
      required:
      - template_name
      type: object
    models.StoreRedeploymentRequest:
      properties:
        data_override:
          additionalProperties:
            type: string
          type: object
        force:
          type: boolean
        shop_id:
          type: string
        subdomain:
          type: string
        template_name:
          type: string
        version:
          type: string
      type: object
...
  securitySchemes:
    OAuth2AccessCode:
      flows:
        authorizationCode:
          authorizationUrl: http://127.0.0.1:8080/oauth2/auth
          tokenUrl: http://127.0.0.1:8080/oauth2/token
      in: header
      type: oauth2
...
info:
  description: This is the Naytife API documentation
  title: Naytife API Docs
  version: "1.0"
openapi: 3.1.0
paths:
...
  /shops/{shop_id}/deploy:
    post:
      description: Deploy a store using the store deployer service
      parameters:
      - description: Shop ID
        in: path
        name: shop_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/models.StoreDeploymentRequest'
        description: Deployment parameters
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.DeploymentResponse'
          description: Store deployed successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Invalid request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Internal server error
      security:
      - OAuth2AccessCode: []
      summary: Deploy a store
      tags:
      - deployment
  /shops/{shop_id}/deployment-status:
    get:
      description: Get the current deployment status of a store
      parameters:
      - description: Shop ID
        in: path
        name: shop_id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.DeploymentStatus'
          description: Deployment status retrieved
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Invalid shop ID
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Internal server error
      security:
      - OAuth2AccessCode: []
      summary: Get deployment status
      tags:
      - deployment
...
  /shops/{shop_id}/redeploy:
    post:
      description: Redeploy an existing store using the store deployer service
      parameters:
      - description: Shop ID
        in: path
        name: shop_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/models.StoreRedeploymentRequest'
        description: Redeployment parameters
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.DeploymentResponse'
          description: Store redeployed successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Invalid request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Internal server error
      security:
      - OAuth2AccessCode: []
      summary: Redeploy a store
      tags:
      - deployment
  /shops/{shop_id}/update-data:
    post:
      description: Update store data without redeploying assets
      parameters:
      - description: Shop ID
        in: path
        name: shop_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/models.DataUpdateRequest'
        description: Data update request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.SuccessResponse'
          description: Store data updated successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Invalid request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Internal server error
      security:
      - OAuth2AccessCode: []
      summary: Update store data
      tags:
      - deployment
...
  /templates:
    get:
      description: Get a list of all available templates from template registry
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/models.Template'
                type: array
          description: Templates retrieved successfully
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Internal server error
      security:
      - OAuth2AccessCode: []
      summary: List all templates
      tags:
      - templates
  /templates/{name}:
    get:
      description: Get details of a specific template by name
      parameters:
      - description: Template name
        in: path
        name: name
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Template'
          description: Template retrieved successfully
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Template not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Internal server error
      security:
      - OAuth2AccessCode: []
      summary: Get a specific template
      tags:
      - templates
  /templates/{name}/latest:
    get:
      description: Get the latest version information for a specific template
      parameters:
      - description: Template name
        in: path
        name: name
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.TemplateVersion'
          description: Latest template version retrieved successfully
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Template not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Internal server error
      security:
      - OAuth2AccessCode: []
      summary: Get latest version of a template
      tags:
      - templates
  /templates/{name}/versions:
    get:
      description: Get a list of all versions for a specific template
      parameters:
      - description: Template name
        in: path
        name: name
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/models.TemplateVersion'
                type: array
          description: Template versions retrieved successfully
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Template not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Internal server error
      security:
      - OAuth2AccessCode: []
      summary: Get all versions of a template
      tags:
      - templates
  /templates/{name}/versions/{version}:
    get:
      description: Get details of a specific template version
      parameters:
      - description: Template name
        in: path
        name: name
        required: true
        schema:
          type: string
      - description: Template version
        in: path
        name: version
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.TemplateVersion'
          description: Template version retrieved successfully
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Template version not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Internal server error
      security:
      - OAuth2AccessCode: []
      summary: Get a specific template version
      tags:
      - templates
  /templates/{name}/versions/{version}/download:
    get:
      description: Download assets for a specific template version
      parameters:
      - description: Template name
        in: path
        name: name
        required: true
        schema:
          type: string
      - description: Template version
        in: path
        name: version
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.TemplateDownload'
          description: Template download information
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Template version not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Internal server error
      security:
      - OAuth2AccessCode: []
      summary: Download a template version
      tags:
      - templates
  /templates/{template_name}/latest:
    get:
      description: Get the latest version for a specific template
      parameters:
      - description: Template name
        in: path
        name: template_name
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                description: Generic API response
                properties:
                  code:
                    example: 200
                    type: integer
                  data: {}
                  message:
                    example: Object updated successfully
                    type: string
                  meta:
                    $ref: '#/components/schemas/models.Meta'
                  status:
                    example: success
                    type: string
                type: object
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Internal Server Error
      security:
      - OAuth2AccessCode: []
      summary: Get latest template version
      tags:
      - templates
...
servers:
- url: http://127.0.0.1:8080/v1